refactor: clean up bad index imports
